requirements.txt
requests
opencv-python
numpy
aiodns
aiohttp